+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## master
+
+* [improvement] `shelly add` should ask about databases instead of database
+* [improvement] Improved messages displayed while deleting the application
+
 ## 0.4.17 / 2013-12-03
 
 * [bugfix] Remove ssh_key from Shelly Cloud only if it was already uploaded